+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Added
+* Added a new method `validateElement` to both do a screenshot comparison and test accessibility
 
 4.25.0 - (March 5, 2019)
 ----------
@@ -13,9 +15,6 @@ Unreleased
 ----------
 ### Changed
 * Updated aggregate translations script fallback to base locale if translation is missing for regional locale on a string-by-string basis.
-
-### Added
-* Added a new method `validateElement` to both do a screenshot comparison and test accessibility
 * Added a script to aggregate dependency themes into a single file
 * Add en-AU to supported translations list